Add Turbo Frames to our articles CRUD
At this stage, flash messages don't work.

When clicking on the "New article" link, here is the full response:

```html
<h1>New article</h1>

<a href="/articles">Back to articles</a>

<turbo-frame id="new_article">
  <form action="/articles" accept-charset="UTF-8" method="post">
    <input type="hidden" name="authenticity_token" value="u9Ai3Rh1F3NMFsXaCPbChndn7Cd0mNAGvQOmXG9XjTgSQUU7w-s3TBLzuIl8KmSBl1J5_G8LcuO7CzZ2VYXq4Q" autocomplete="off" />
    <label for="article_content">Content</label>
    <input type="text" name="article[content]" id="article_content" />

    <input type="submit" name="commit" value="Create Article" data-disable-with="Create Article" />
  </form>
</turbo-frame>
```

As we can see, we don't have a layout, and this is expected behavior
because of the `Turbo::Frames::FrameRequest` concern:

```rb
layout -> { false if turbo_frame_request? }
```

Then, when we click on the "Create Article" button, we don't have a
layout either by default:

```html
<turbo-stream action="update" target="new_article"><template></template></turbo-stream>
<turbo-stream action="append" target="articles"><template><div>
  Article #8:
</div>
</template></turbo-stream>
```

This is what we will try to change in the next commit.