Enable flash messages for Turbo Stream requests
In this commit, we tell Rails to use the `application.html.erb` layout
for HTML requests and the layout `application.turbo_stream.erb` for
Turbo Stream requests.

Flash messages now work as expected when we click on the
"Create Article" button, but there is an unintended side effect. When
clicking on the "New article" link, the layout is now included in the
response:

```html
<!DOCTYPE html>
<html>
  <head>
    ...
  </head>

  <body>
    <div id="flash"></div>

    <h1>New article</h1>

    <a href="/articles">Back to articles</a>

    <turbo-frame id="new_article">
      <form action="/articles" accept-charset="UTF-8" method="post">
        ...
      </form>
    </turbo-frame>
  </body>
</html>
```

This is not what we wanted. What we really want to do is to remove the
layout for Turbo Frame requests that are not Turbo Stream requests.

We'll fix this in the next commit.